+++ tests/test_device.py
@@ -35,7 +35,7 @@ async def test_device_general(factory: helper.Factory) -> None:
     assert device.interface_id == const.INTERFACE_ID
     assert device.has_custom_entity_definition is True
     assert len(device.custom_entities) == 3
-    assert len(device.channels) == 10
+    assert len(device.channels) == 11
 
 
 @pytest.mark.asyncio